Add IF::View role to abstract display of info
The Engine has a View, and registers event listeners that use the View
to display info to the player. For example, a describe-room event will
fire a listener that shows the room description.

Stubs for several views exist: Null (display nothing), Std (print msg
to $*OUT), and Test (validates that the expected info is being
displayed).

This commit also enables multiple listeners for the same event.