Add an EOF event when event queue is empty
Rearranges the event processing so that events that are emitted from a
listener don't interrupt that listener. They get handled afterwards.

The handling of the EOF event is tricky because an EOF listener can emit
events, too. Those events must be processed, but not trigger another EOF
event.

The EOF event is used to display a prompt for input on the view.