Enable input via IF::View and events
Make IF::View a class instead of a role, to be able have the $!events
attribute. View's BUILD registers listeners for view-related stuff. This
means the IF::Engine doesn't need to know about the view. It relies only
on events for communicating.

The View holds a handle to the engine's do() function, to send input to
the game engine model.

The EOF event handling is improved, so it's possible to use it to create
the event loop by just emitting an event from an EOF listener.

It is possible now to play bin/crypt.pl in a rudimentary fashion.