View::Test: easier testing, can skip some actions
Rename verify() to verify-all(), and have verify() silently skip
non-matching actions. Just test the stuff that really matters, and
ignore the details.

Also generalizes the tests so that instead of forcing an exact match on
a string, can pass a regex or code block and test on those. This pulls
the test that was in TestOut class up into the IF::Test module.

Other minor test cleanups.