Move between rooms
Just make the directions strings instead of an enum. Too much
hassle dealing with the enum.

Use the / <{ join '|', @foo }> / trick to match one of a list. I'm
not able to get a cleaner syntax to work in the regex.

As an aside, it's fun to walk back and forth in the game.

Warning, the rooms refer to each other, and .perl doesn't handle
cycles in the data structures. Both Rakudo and Niecza spin forever
on that. So don't use .perl on rooms.

Moved setting $!room from 'begin' to 'enter-room' listener. It
is pretty nice.